test: update MCP tests for refactored core
